updating code

+++ inc/Grid.hh
@@ -8,7 +8,7 @@ class Grid
 public:
   Grid();
   ~Grid();
-  int size() const { return _x.size(); }
+  int size() const { return (int) _x.size(); }
   double at(int i) const { return _x[i]; }
 private:
   vector<double> _x;
